[C10D] Decouple PGNCCL desync from dbg dump
Add TORCH_NCCL_DUMP_DEBUG_INFO env to control dumping independently
of desync debug feature.

Currently default to disabled (so no behavior change by default),
but plan to default this to true after validation.

Keeps the 'sleep for 30 sec' that used to be after desync debug where it
was.

Fixes #114433

ghstack-source-id: eb3f819c1640fe9daba3b8fd849173d00bcf9696
Pull Request resolved: https://github.com/pytorch/pytorch/pull/114614