[C10D] Add time_created_us to flight recorder
time_created_us is the cpu-side epoch_time (in usec) when a flight-recorder
event was created. It loosely corresponds to the time the c10d collective
API was called and a work object was created.  It does NOT correspond to
the time the collective started on the GPU.

We follow the precedent of us epoch time from this PR adding timestamps
to the cuda caching allocator:
https://github.com/pytorch/pytorch/pull/112266

ghstack-source-id: e2f3850b5ac57ae00b4f42aa82c079d2f3c2a1d2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/114810